catch trigger client exception

+++ tests/S3RestoreTest.php
@@ -560,7 +560,7 @@ class S3RestoreTest extends BaseTest
             'id' => [
                 'KBC.datatype.type' => 'NUMBER',
                 'KBC.datatype.nullable' => '',
-                'KBC.datatype.basetype' => 'NUMERIC',
+                'KBC.datatype.basetype' => 'INTEGER',
                 'KBC.datatype.length' => '38,0',
             ],
             'name' => [
